chore(editor): extract binary asset helpers
